Break circular import in reactions model
